Função SalvarCalculo e CarregarCalculo atualizada

+++ PI2/CalculoSuspensao.cs
@@ -8,14 +8,14 @@ namespace PI2
 {
     public static class CalculoSuspensao
     {
-        public static float CalcularConstanteElastica(float f, float bd, float sd)
+        public static double CalcularConstanteElastica(double f, double bd, double sd)
         {
             if (sd == 0)
             {
                 throw new DivideByZeroException("Não dá pra dividir por 0.");
             }
 
-            float resultado_em_kg_mm, resultado_em_N_m;
+            double resultado_em_kg_mm, resultado_em_N_m;
 
             resultado_em_kg_mm = (f * bd) / sd;
             resultado_em_N_m = resultado_em_kg_mm * 10 * 1000;
@@ -23,36 +23,36 @@ namespace PI2
             return resultado_em_N_m;
         }
 
-        public static float CalcularCursoAngular(float a, float b)
+        public static double CalcularCursoAngular(double a, double b)
         {
             if (b == 0)
             {
                 throw new DivideByZeroException("Não dá pra dividir por 0.");
             }
 
-            float resultado;
+            double resultado;
 
-            resultado = (float) (2 * Math.Asin(a / b));
+            resultado = 2 * Math.Asin(a / b);
 
             resultado = RadianosParaGraus(resultado);
 
             return resultado;
         }
 
-        public static float RadianosParaGraus(float a)
+        public static double RadianosParaGraus(double a)
         {
-            float resultado;
+            double resultado;
 
-            resultado = (float) (a * 180 / Math.PI);
+            resultado = a * 180 / Math.PI;
 
             return resultado;
         }
 
-        public static float GrausParaRadianos(float a)
+        public static double GrausParaRadianos(double a)
         {
-            float resultado;
+            double resultado;
 
-            resultado = (float) (a * Math.PI / 180);
+            resultado = a * Math.PI / 180;
 
             return resultado;
         }
